requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Limiter==3.12
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
WTForms==3.2.1
alembic==1.14.1
email-validator==2.2.0
fpdf2==2.8.2
openpyxl==3.1.5
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
gunicorn==23.0.0
twilio==9.4.1
pytest==8.3.4
flake8==7.1.1
